site/requirements.txt
beautifulsoup4==4.12.3
jinja2==3.1.4
pillow==10.4.0
PyYAML==6.0.2
tqdm==4.66.5
